common: double width of sUnassociatedStationStats.time_stamp to delay overflow
EasyMesh Unassociated STA Link Metrics Response TLV contains
a 4byte / 32bit field that should hold the time delta
between the 'time point when the measurement was made' and
the 'time point when the Report was sent'

PWHM exposes time as a RFC3339 string
using lib amxc to parse this string into a struct containing
a signed 64bit 'seconds' field.

https://prplfoundationcloud.atlassian.net/browse/PPM-3413

+++ common/beerocks/bwl/dummy/mon_wlan_hal_dummy.cpp
@@ -487,7 +487,7 @@ bool mon_wlan_hal_dummy::sta_unassoc_rssi_measurement(
         100, //signal_strength
         36,  //channel
         115, //operating_class(5Ghz)
-        (uint32_t)now_c,
+        (uint64_t)now_c,
     };
 
     stats.push_back(new_stat);
@@ -497,7 +497,7 @@ bool mon_wlan_hal_dummy::sta_unassoc_rssi_measurement(
         200, //signal_strength
         1,   //channel
         81,  //operating_class(2Ghz)
-        (uint32_t)now_c,
+        (uint64_t)now_c,
     };
     stats.push_back(new_stat);
